+++ RxAlamofire/Pods/Target Support Files/Alamofire-iOS9.0/Alamofire-iOS9.0-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_Alamofire_iOS9_0 : NSObject
+@end
+@implementation PodsDummy_Alamofire_iOS9_0
+@end