+++ RxAlamofire/Pods/Target Support Files/Alamofire-tvOS9.0/Alamofire-tvOS9.0-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_Alamofire_tvOS9_0 : NSObject
+@end
+@implementation PodsDummy_Alamofire_tvOS9_0
+@end