+++ RxAlamofire/Pods/Target Support Files/RxCocoa-watchOS3.0/RxCocoa-watchOS3.0-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_RxCocoa_watchOS3_0 : NSObject
+@end
+@implementation PodsDummy_RxCocoa_watchOS3_0
+@end