+++ RxAlamofire/Pods/Target Support Files/RxSwift-watchOS3.0/RxSwift-watchOS3.0-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_RxSwift_watchOS3_0 : NSObject
+@end
+@implementation PodsDummy_RxSwift_watchOS3_0
+@end