+++ RxAlamofire/RxAlamofire-watchOS/RxAlamofire_watchOS.h
@@ -1,0 +1,19 @@
+//
+//  RxAlamofire_watchOS.h
+//  RxAlamofire-watchOS
+//
+//  Created by Ryne Cheow on 19/4/19.
+//  Copyright © 2019 Bonto.ch. All rights reserved.
+//
+
+#import <WatchKit/WatchKit.h>
+
+//! Project version number for RxAlamofire_watchOS.
+FOUNDATION_EXPORT double RxAlamofire_watchOSVersionNumber;
+
+//! Project version string for RxAlamofire_watchOS.
+FOUNDATION_EXPORT const unsigned char RxAlamofire_watchOSVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <RxAlamofire_watchOS/PublicHeader.h>
+
+